Merged /branches/cleansim r1909-1937
Included Eurocontrol's public aircraft perfomance tables
http://www.eurocontrol.be/eec/public/standard_page/2004_note_12.html so that aircraft behave much more realistically while climbimg and descending.

Together with the performance tables the documentation gave all the formulas to correctly calculate atmosphere parameters and tas/cas conversions, and created a new Atmosphere class.

Updated Modelos_avo.txt so that it has an equivalence list to the aircraft models available in BADA

Removed some of the cruft of old code.